+++ crates/objc2/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
+### Changed
+* Moved `ClassBuilder` and `ProtocolBuilder` from the `declare` module to the
+  `runtime` module. The old locations are deprecated.
+
 
 ## 0.5.0 - 2023-12-03